# backend/internal/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_studentinovisad_popisomator_backend_internal_dto.CreatePropertyRequest:
    properties:
      default_value:
        type: string
      description:
        type: string
      name:
        type: string
      value_type:
        enum:
        - string
        - number
        - boolean
        type: string
    required:
    - name
    - value_type
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - manager
        - user
        type: string
      status:
        enum:
        - requested
        - active
        type: string
    required:
    - email
    - full_name
    - password
    - role
    - status
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.PropertiesPage:
    properties:
      items:
        description: Property page items
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Property'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.Property:
    properties:
      default_value:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      value_type:
        type: string
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.PropertyOption:
    properties:
      id:
        type: integer
      name:
        type: string
      value_type:
        type: string
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.UpdatePropertyRequest:
    properties:
      default_value:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.UpdateUserRequest:
    properties:
      role:
        enum:
        - admin
        - manager
        - user
        type: string
      status:
        enum:
        - active
        - requested
        type: string
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.User:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_repository.UserRole'
      status:
        $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_repository.UserStatus'
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.UsersPage:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.User'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_repository.UserRole:
    enum:
    - admin
    - manager
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleManager
    - UserRoleUser
  github_com_studentinovisad_popisomator_backend_internal_repository.UserStatus:
    enum:
    - requested
    - active
    type: string
    x-enum-varnames:
    - UserStatusRequested
    - UserStatusActive
  github_com_studentinovisad_popisomator_backend_internal_response.Error:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Inventory management API for Popisomator.
  title: Popisomator API
  version: "1.0"
paths:
...
  /properties:
    get:
      parameters:
      - description: Page size (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.PropertiesPage'
        "400":
          description: invalid limit/offset
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List property definitions
      tags:
      - Properties
    post:
      consumes:
      - application/json
      parameters:
      - description: Property to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.CreatePropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Property'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Create a property definition (manager/admin only)
      tags:
      - Properties
  /properties/{id}:
    delete:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid property id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Delete a property definition (manager/admin only)
      tags:
      - Properties
    get:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Property'
        "400":
          description: invalid property id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Get a property definition by ID
      tags:
      - Properties
    patch:
      consumes:
      - application/json
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UpdatePropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.Property'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update a property definition (manager/admin only)
      tags:
      - Properties
  /properties/options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.PropertyOption'
            type: array
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List all properties in a minified form, for dropdowns
      tags:
      - Properties
  /users:
    get:
      parameters:
      - description: Page size (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by name/email substring (max 100 chars)
        in: query
        name: search
        type: string
      - description: Filter by role
        enum:
        - admin
        - manager
        - user
        in: query
        name: role
        type: string
      - description: Filter by status
        enum:
        - requested
        - active
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UsersPage'
        "400":
          description: invalid limit/offset/search/role/status
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List users (admin only)
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: User to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.User'
        "400":
          description: invalid request / validation failed
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "409":
          description: user with this email already exists
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Create a new user (admin only, created active)
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid user id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Delete a user (admin only)
      tags:
      - Users
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.User'
        "400":
          description: invalid request / validation failed
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update a user's role/status (admin only)
      tags:
      - Users
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.User'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Get the currently authenticated user's own details
      tags:
      - Users
securityDefinitions:
  CookieAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"